requirements.txt
langchain
langchain-community
langchain-experimental
langchain-huggingface
langchain-groq
sentence-transformers
faiss-cpu
pypdf
python-pptx
python-docx
pandas
openpyxl
PyMuPDF
pdfplumber
gradio
python-dotenv
